broke out constants and common.
